+++ .run/build.run.xml
@@ -1,0 +1,12 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="build" type="js.build_tools.npm" nameIsGenerated="true">
+    <package-json value="$PROJECT_DIR$/package.json" />
+    <command value="run" />
+    <scripts>
+      <script value="build" />
+    </scripts>
+    <node-interpreter value="$USER_HOME$/.nvm/versions/node/v18.0.0/bin/node" />
+    <envs />
+    <method v="2" />
+  </configuration>
+</component>
