+++ examples/make_image.js
@@ -1,4 +1,4 @@
-import * as PImage from "../../src/index.js"
+import * as PImage from "../dist/index.ems.js"
 import * as fs from 'fs'
 const img1 = PImage.make(100, 100)
 const ctx = img1.getContext('2d');
@@ -8,5 +8,5 @@ ctx.fillRect(0,0,100,100);
 PImage.encodePNGToStream(img1, fs.createWriteStream('out.png')).then(() => {
     console.log("wrote out the png file to out.png");
 }).catch((e)=>{
-    console.log("there was an error writing");
+    console.log("there was an error writing",e);
 });
