+++ examples/stream_promise.js
@@ -4,13 +4,13 @@ draws the current date in big black letters,
 and writes the final image to disk
 
  */
-import * as PImage from "../../src/index.js"
+import * as PImage from "../dist/index.ems.js"
 import fs from 'fs'
 import * as client from "https"
 
 let url = "https://vr.josh.earth/webxr-experiments/physics/jinglesmash.thumbnail.png"
-let filepath = "output.png"
-let fontpath = 'test/unit/fixtures/fonts/SourceSansPro-Regular.ttf'
+let filepath = "output_stream_promise.png"
+let fontpath = '../test/unit/fixtures/fonts/SourceSansPro-Regular.ttf'
 PImage.registerFont(fontpath,'MyFont').loadPromise()
     //Promise hack because https module doesn't support promises)
     .then(()=>new Promise(res => client.get(url,res)))
